requirements.txt
numpy==1.22.3
Pillow==9.1.1
qoi==0.0.9
